Implemented type 1 block transfers
First and foremost, this is not a well made implementation of block1 transfers.
It appears to work, and limited testing has been performed against the a Java implementation of CoAP, specifically Californium.
It should support the server requesting a smaller block size, however this hasn't been tested.

Other fun issues include:
Extensive testing has not been performed
Code does not conform to the repository conventions
Observe functionality has not been considered
There's a hack on line 511 in server.js to ensure responses from the server after a block1 transfer is complete do not cause issues on the client side
And possibly more that I don't remember.

This was a rushed implementation which will probably be revised in the future.
Use at your own risk. :boom: