test: change tests for block2 multicast
Remove the second server from the two blockwise multicast test cases
as the changes from the commit before cause the agent to ignore
responses that are sent twice -- as both servers in the two test cases
used the same IP address in their responses the tests timed out.
The tests, however, should still be valid as the server still responds
despite a changed IP address beginning with the second request.