WSHUB-360: request tests: Wrap call-back in try-catch.
If an exception is thrown in the callback, there is no guarantee that
the test case will "see" it due to the function being called
asynchronously.  Better bet: use try-catch inside the callback, then we
can pass that to `done` so the exception is caught and processed by the
test suite.