Add support for de-registering observations (#214)
* Allow url.observe to be a number

RFC 7641 Section 2 defined that the Observe option may have an integer value:
   0 (register) adds the entry to the list, if not present;
   1 (deregister) removes the entry from the list, if present.

This commit allows the url.observe to be also a JS number object,
instead of just `true`, thereby adding support for observe deregistration.

* WSHUB-360: tests: Add test case for sending Observe=1

As per advice in https://github.com/mcollina/node-coap/pull/164.  This
bug has sat there unaddressed long enough!

* WSHUB-360: observe_read_stream: Add `eager` parameter to `close()`.

If truthy, we emit a `deregister` signal so the agent can do its bit.
(2021-06-12: Rename parameter as per @JKRhb's suggestion.)

* WSHUB-360: agent: Send deregister if requested.

If the calling application calls `.close(true)`, then the agent should
detect the `deregister` signal emitted and submit a request to the CoAP
server de-registering it from the subscription.

* WSHUB-360: proxy tests: Set timeout, catch errors

- This seems to want to hang for about 15 seconds… possibly a delay in
  looking up the bad hostname.  So give the test a bit more time.
- When we do get a response, catch the errors and pass them to the
  `done` callback.  Don't let them bubble up to the interpreter as that
  will cause the test to "time out" instead of failing.

* WSHUB-360: request tests: extend timeouts

Looking up `aaa.eee` may take longer than 2 seconds.

* WSHUB-360: server: Catch and forward errors.

If binding fails at this point, exceptions can leak which leads to test
errors like this:

```
    multicast
Caught exception: Error: addMembership EADDRINUSE
Caught exception: Error: addMembership EADDRINUSE
      1) receive CoAp message
      ✓ receive CoAp message
      2) "after each" hook for "receive CoAp message"

  validate custom server options

  294 passing (32s)
  2 failing

  1) server multicast receive CoAp message:
     Uncaught Error: addMembership EADDRINUSE
      at Socket.addMembership (dgram.js:833:11)
      at lib/server.js:234:26
      at Array.forEach (<anonymous>)
      at Socket.<anonymous> (lib/server.js:233:44)
      at Socket.onListening (dgram.js:228:10)
      at startListening (dgram.js:153:10)
      at dgram.js:346:7
      at processTicksAndRejections (internal/process/task_queues.js:85:21)

  2) server "after each" hook for "receive CoAp message":
     Error [ERR_SOCKET_DGRAM_NOT_RUNNING]: Not running
      at healthCheck (dgram.js:902:11)
      at Socket.close (dgram.js:724:3)
      at Context.<anonymous> (test/server.js:42:12)
      at processImmediate (internal/timers.js:456:21)
```

Much better, if we use a `try`-`catch` to capture those, so the `done`
callback can be used to forward them.

* WSHUB-360: agent: Use hex string token as key.

We need to reference a particular request by its token.  The old way
assumed it was a 32-bit integer, which works, but is inflexible.  If we
just encode it as hexadecimal, it'll make the property names a bit
shorter whilst removing a hard dependency on 4-byte tokens.

* WSHUB-360: agent: Allow user to specify their own token.

* WSHUB-360: agent: Validate token length.

Standards dictate it shall not be longer than 8 bytes.
(Whilst we're here, also check for 0-length token.)

* WSHUB-360: request tests: Test custom token.

* WSHUB-360: request tests: Skip 'ACK' in de-register test

Also catch the exceptions that take place in the call-back function.

* WSHUB-360: server tests: Catch expect errors

Forward these to `done` as expected by the test harness.

* WSHUB-360: agent: Clean up missed ._nextToken() call

* WSHUB-360: request tests: Wrap call-back in try-catch.

If an exception is thrown in the callback, there is no guarantee that
the test case will "see" it due to the function being called
asynchronously.  Better bet: use try-catch inside the callback, then we
can pass that to `done` so the exception is caught and processed by the
test suite.

* WSHUB-360: request test cases: Fix `let` slipping out.

`node-coap` uses an older JavaScript revision than I'm used to.

* WSHUB-360: request tests: Test empty token case

* WSHUB-360: agent: Make token length 8 bytes.

We'll add some pseudorandom bytes to the end so that the token remains
unique.

Co-authored-by: Pekka Nikander <pekka.nikander@iki.fi>