Update automatic blockwise transfer logic
Fix for blockwise transfer issues raised in #373 where a coap message
could avoid being split into blocks, yet still be larger than the
allowable size of 1280 bytes. This applies the recommended maximum
payload size from RFC 7252 Section 4.6 of 1024 bytes, and allows it to
be reduced in the config parameters as required.

The error that this fixes was introduced in
d727d430d7c744a516d4c68d156ed6bf8262f7b7 and this attempts to keep the
intent behind that commit, whilst fixing the technical flaws that it
introduced.

The 'IP_MTU' constant has been renamed to 'MAX_PAYLOAD' to more
accurately describe what the it represents. The 'maxPacketSize'
parameter has also been renamed to 'maxPayloadSize' for the same
reason.

Note that the renaming of the parameter is a potentially breaking
change.